Minor feedback points from code review
